# api/openapi.yaml
openapi: 3.0.3
info:
  title: corps (团队) API
  version: 0.6.0
  description: >
    MVP API for the corps multi-tenant collaboration SaaS. This document is the
    single contract for frontend and backend in Phase 3. All paths are prefixed with
    /api/v1/. Every response uses the uniform envelope { code, data, message }.
    Multi-tenant access control: every {wid} route verifies the caller's membership
    row for that workspace, then runs all tenant queries inside a transaction with
    app.workspace_id / app.user_id set (RLS binds at engine level in hardened
    deployments — see db/schema.sql).
    This contract currently defines 57 paths (updated 2026-09-09).
  contact:
    name: MVP Dev Expert Team — Architecture
servers:
  - url: http://localhost:3000
    description: Local development (Next.js Route Handlers)
  - url: https://ghcr.io/{env}/corps
    description: GHCR Docker deployment
    variables:
      env:
        default: staging
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/workspaces/{wid}/members/{userId}:
    patch:
      tags: [members]
      summary: Change a member's role (owner/admin)
      description: >
        Owner/Admin only (member → 403). Allowed target roles are `admin` / `member`;
        the workspace owner's role cannot be changed (403). Unknown member → 404.
      operationId: updateMemberRole
      parameters:
        - $ref: '#/components/parameters/Wid'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, member]
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer, example: 200 }
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      role: { type: string, enum: [admin, member] }
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [members]
      summary: Remove a member
      description: Owner/Admin only. Removing the last owner is rejected (400).
      operationId: removeMember
      parameters:
        - $ref: '#/components/parameters/Wid'
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Null'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # F2: Workspace permission matrix (role-based R/C/U/D overrides)
  /api/v1/workspaces/{wid}/permissions:
    get:
      tags: [members]
      summary: Get workspace permission matrix
      description: >
        Returns the merged permission matrix for all roles (default + overrides).
        The matrix maps each role (owner/admin/member/viewer) to a set of
        module→actions entries. Owner row is always full and locked; Admin can
        override Member and Viewer columns.
      operationId: getPermissionMatrix
      parameters:
        - $ref: '#/components/parameters/Wid'
      responses:
        '200':
          description: Permission matrix
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_PermissionMatrix'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [members]
      summary: Update permission override
      description: >
        Update permission override for a specific role and module. Only `member`
        and `viewer` roles can be overridden; attempting to override `owner` or
        `admin` returns 400. Admin/Owner only (member → 403).
      operationId: updatePermissionOverride
      parameters:
        - $ref: '#/components/parameters/Wid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePermissionRequest'
      responses:
        '200':
          description: Permission override updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_PermissionMatrix'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # INVITATIONS（邀请未注册用户）
  # ===========================================================================
...
  /api/v1/workspaces/{wid}/billing/portal:
    post:
      tags: [billing]
      summary: Create a Stripe Customer Portal session
      description: >
        Owner only. Returns a Stripe Billing Portal URL so the customer can update payment
        method / cancel. Requires an existing Stripe customer (from a completed checkout).
      operationId: billingPortal
      parameters:
        - $ref: '#/components/parameters/Wid'
      responses:
        '200':
          description: Portal session URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Checkout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
        '501':
          $ref: '#/components/responses/PaymentNotConfigured'
...
components:
  securitySchemes:
    cronSecret:
      type: http
      scheme: bearer
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access JWT (15 min) in Authorization header. workspace_id resolved from JWT claim `wid`.
    refreshAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: >
        Better Auth session cookie (httpOnly, secure, 7-day). The /api/v1/auth/refresh
        endpoint reads this cookie to re-issue a workspace-scoped access JWT.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: HMAC-SHA256 signature of the raw body using the Stripe webhook secret.
  parameters:
    Wid:
      name: wid
      in: path
      required: true
      description: Workspace id. Must equal the JWT `wid` claim; mismatch is 403/404.
      schema: { type: string, format: uuid }
...
    UserId:
      name: userId
      in: path
      required: true
      description: User id (UUID) of the member to update or remove.
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Validation error (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 400
            data: null
            message: "email must be a valid address"
    Unauthorized:
      description: Missing or invalid token (401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 401
            data: null
            message: "invalid or expired access token"
    Forbidden:
      description: Authenticated but not permitted (RBAC / tenant mismatch) (403)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 403
            data: null
            message: "role 'member' cannot perform this action"
    NotFound:
      description: Resource unknown to caller (RLS-filtered) (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 404
            data: null
            message: "workspace not found"
...
    PaymentNotConfigured:
      description: Payment provider not configured (501)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 501
            data: null
            message: "payment provider not configured"
    InternalError:
      description: Internal server error (500)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            code: 500
            data: null
            message: "internal server error"
...
  schemas:
...
    Envelope:
      type: object
      required: [code]
      properties:
        code:
          type: integer
          description: "Business status code, mirrors HTTP status code (200=success, 4xx/5xx=error)"
          example: 200
        data:
          nullable: true
        message:
          type: string
          description: "Human-readable status message. Required for error responses (4xx/5xx), optional for success responses (2xx)."
          example: "ok"
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 401
        data:
          nullable: true
          example: null
        message:
          type: string
          example: "invalid or expired access token"
        # L4：events POST 400 等校验失败响应会附带 zod errors 数组，
        # 此前未在 ErrorEnvelope 中文档化。这里补为可选字段（未列入 required），
        # 既覆盖现有带 errors 的 4xx 响应，也不破坏只返回 {code,message,data} 的错误响应。
        errors:
          type: array
          description: >
            Optional field-level validation errors (e.g. zod issues) for 4xx
            responses. Absent on non-validation errors.
          items:
            type: object
            description: A single validation issue (path/message/code as produced by zod).
    Envelope_Null:
      allOf:
        - $ref: '#/components/schemas/Envelope'
      properties:
        data:
          nullable: true
    # L2：Stripe webhook 200 实际返回 { received: true } 或 { received: true, duplicate: true }，
    # 此前误用 Envelope_Null（data:null），与实现不一致。这里定义专用 envelope，
    # data 形状为 { received: boolean, duplicate?: boolean }，与 route.ts 两处 200 应答对齐。
...
    Envelope_Checkout:
      allOf:
        - $ref: '#/components/schemas/Envelope'
      properties:
        data:
          $ref: '#/components/schemas/CheckoutSession'
...
    CheckoutSession:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: Redirect URL for Stripe Checkout / Alipay gateway / Billing Portal. Present for stripe and alipay-page channels; absent for wechatpay-native (use qrCodeUrl instead).
        qrCodeUrl:
          type: string
          description: WeChat Pay code_url for QR rendering. Present for the wechatpay-native channel; absent for stripe and alipay-page.
        providerOrderId:
          type: string
          description: Channel-side order/session identifier created by the payment provider.
        providerId:
          type: string
          enum: [stripe, wechatpay-native, alipay-page]
          description: Payment channel that handled this checkout session.
...
    PermissionOverride:
      type: object
      required: [role, module, actions]
      properties:
        role:
          type: string
          enum: [member, viewer]
          description: Role whose permissions are overridden. Owner/admin cannot be overridden.
        module:
          type: string
          description: Module key (e.g. tasks, decisions, documents, members, billing, analytics, settings).
          example: tasks
        actions:
          type: array
          items:
            type: string
            enum: [read, create, update, delete]
          description: Allowed actions for this role+module. Absent actions are denied.
          example: [read]
    PermissionMatrix:
      type: object
      required: [roles, overrides]
      properties:
        roles:
          type: array
          description: Ordered list of role permission rows (owner, admin, member, viewer).
          items:
            type: object
            required: [role, modules]
            properties:
              role:
                type: string
                enum: [owner, admin, member, viewer]
              locked:
                type: boolean
                description: true when this role row cannot be modified (owner is always locked).
              modules:
                type: array
                items:
                  type: object
                  required: [module, actions]
                  properties:
                    module:
                      type: string
                      description: Module key.
                    actions:
                      type: array
                      items:
                        type: string
                        enum: [read, create, update, delete]
                      description: Allowed actions for this role+module.
        overrides:
          type: array
          description: Active permission overrides (only member/viewer rows can have overrides).
          items:
            $ref: '#/components/schemas/PermissionOverride'
    Envelope_PermissionMatrix:
      allOf:
        - $ref: '#/components/schemas/Envelope'
      properties:
        data:
          $ref: '#/components/schemas/PermissionMatrix'
    UpdatePermissionRequest:
      type: object
      required: [role, module, actions]
      properties:
        role:
          type: string
          enum: [member, viewer]
          description: Target role to override. Owner/admin are rejected with 400.
        module:
          type: string
          minLength: 1
          maxLength: 50
          description: Module key (e.g. tasks, decisions, documents).
        actions:
          type: array
          items:
            type: string
            enum: [read, create, update, delete]
          description: Allowed actions. Empty array denies all actions for this role+module.